Crate kabuki

Actor library built on top of the Tokio platform.

Kabuki provides a simple way to structure concurrent applications built for the Tokio / Futures platform. It is based on the actor model. An actor is a small unit of computation that is used to manage state and resources. It receives messages from other actors and performs some kind of action based on that input. This way, instead of having state and resources accessed concurrently, only a single thread access it and concurrent access is handled through message passing.

Motivation

Tokio and Futures provide a lightweight task primitive. However, it leaves the details of managing concurrency up to the developer.

Structs

ActorCell
Manages the runtime state of an Actor.

ActorFn
An actor implemented by a closure.

ActorFuture
The return value of ActorRef::call

ActorRef
Handle to an actor, used to send messages

Builder
Builds an actor

Enums

ActorState
Tracks the state of the actor

CallError
Used to represent call errors

Traits

Actor
A value that manages state and responds to messages

Functions

actor_fn
Returns an Actor backed by the given closure.